double to float

+++ optimizations/JPEG/DCT.cs
@@ -5,25 +5,25 @@ namespace JPEG;
 
 public class DCT
 {
-    private static readonly double[,] CosTable = new double[8, 8];
+    private static readonly float[,] CosTable = new float[8, 8];
 
     static DCT()
     {
         for (var i = 0; i < 8; i++)
         for (var j = 0; j < 8; j++)
-            CosTable[i, j] = Math.Cos(((2 * j + 1) * i * Math.PI) / 16);
+            CosTable[i, j] = (float)Math.Cos(((2 * j + 1) * i * Math.PI) / 16);
     }
 
-    public static double[,] DCT2D(double[,] input)
+    public static float[,] DCT2D(float[,] input)
     {
         var height = input.GetLength(0);
         var width = input.GetLength(1);
-        var coeffs = new double[width, height];
+        var coeffs = new float[width, height];
         var beta = Beta(height, width);
-        var alpha = 1 / Math.Sqrt(2);
+        var alpha = 1 / (float)Math.Sqrt(2);
         
         //u==0 v==0
-        var sum = 0d;
+        var sum = 0f;
         for (var x = 0; x < width; x++)
         {
             for (var y = 0; y < height; y++)
@@ -36,7 +36,7 @@ public class DCT
         //u==0 v!=0
         for (var v = 1; v < height; v++)
         {
-            sum = 0d;
+            sum = 0f;
             for (var x = 0; x < width; x++)
             {
                 for (var y = 0; y < height; y++)
@@ -50,7 +50,7 @@ public class DCT
         //u!=0 v==0
         for (var u = 1; u < width; u++)
         {
-            sum = 0d;
+            sum = 0f;
             for (var x = 0; x < width; x++)
             {
                 for (var y = 0; y < height; y++)
@@ -66,7 +66,7 @@ public class DCT
         {
             for (var v = 1; v < height; v++)
             {
-                sum = 0d;
+                sum = 0f;
                 for (var x = 0; x < width; x++)
                 {
                     for (var y = 0; y < height; y++)
@@ -80,18 +80,18 @@ public class DCT
         return coeffs;
     }
 
-    public static void IDCT2D(double[,] coeffs, double[,] output)
+    public static void IDCT2D(float[,] coeffs, float[,] output)
     {
         var height = coeffs.GetLength(0);
         var width = coeffs.GetLength(1);
         var beta = Beta(height, width);
-        var alpha = 1 / Math.Sqrt(2);
+        var alpha = 1 / (float)Math.Sqrt(2);
         
         for (var x = 0; x <width; x++)
         {
             for (var y = 0; y < height; y++)
             {
-                var sum = 0d;
+                var sum = 0f;
                 
                 //u==0 v==0
                 sum += coeffs[0, 0] * CosTable[0, x] * CosTable[0, y] * alpha * alpha; 
@@ -119,8 +119,8 @@ public class DCT
         }
     }
 
-    private static double Beta(int height, int width)
+    private static float Beta(int height, int width)
     {
-        return 1d / width + 1d / height;
+        return 1f / width + 1f/ height;
     }
 }
