+++ optimizations/JPEG/Processor/JpegProcessor.cs
@@ -15,7 +15,7 @@ public class JpegProcessor : IJpegProcessor
 	private const int DCTSize = 8;
 	private static int[,] _quantizationMatrix;
 	private static byte[,] _quantizedBuffer = new byte[DCTSize, DCTSize];
-	private static double[,] _deQuantizedBuffer = new double[DCTSize, DCTSize];
+	private static float[,] _deQuantizedBuffer = new float[DCTSize, DCTSize];
 
 	static JpegProcessor()
 	{
@@ -48,7 +48,7 @@ public class JpegProcessor : IJpegProcessor
 		{
 			for (var x = 0; x < matrix.Width; x += DCTSize)
 			{
-				foreach (var selector in new Func<Pixel, double>[] { p => p.Y, p => p.Cb, p => p.Cr })
+				foreach (var selector in new Func<Pixel, float>[] { p => p.Y, p => p.Cb, p => p.Cr })
 				{
 					var subMatrix = GetSubMatrix(matrix, y, DCTSize, x, DCTSize, selector);
 					ShiftMatrixValues(subMatrix, -128);
@@ -81,9 +81,9 @@ public class JpegProcessor : IJpegProcessor
 			{
 				for (var x = 0; x < image.Width; x += DCTSize)
 				{
-					var _y = new double[DCTSize, DCTSize];
-					var cb = new double[DCTSize, DCTSize];
-					var cr = new double[DCTSize, DCTSize];
+					var _y = new float[DCTSize, DCTSize];
+					var cb = new float[DCTSize, DCTSize];
+					var cr = new float[DCTSize, DCTSize];
 					foreach (var channel in new[] { _y, cb, cr })
 					{
 						var quantizedBytes = new byte[DCTSize * DCTSize];
@@ -102,7 +102,7 @@ public class JpegProcessor : IJpegProcessor
 		return result;
 	}
 
-	private static void ShiftMatrixValues(double[,] subMatrix, int shiftValue)
+	private static void ShiftMatrixValues(float[,] subMatrix, int shiftValue)
 	{
 		var height = subMatrix.GetLength(0);
 		var width = subMatrix.GetLength(1);
@@ -112,7 +112,7 @@ public class JpegProcessor : IJpegProcessor
 			subMatrix[y, x] = subMatrix[y, x] + shiftValue;
 	}
 
-	private static void SetPixels(Matrix matrix, double[,] a, double[,] b, double[,] c, PixelFormat format,
+	private static void SetPixels(Matrix matrix, float[,] a, float[,] b, float[,] c, PixelFormat format,
 		int yOffset, int xOffset)
 	{
 		var height = a.GetLength(0);
@@ -123,10 +123,10 @@ public class JpegProcessor : IJpegProcessor
 			matrix.Pixels[yOffset + y, xOffset + x] = new Pixel(a[y, x], b[y, x], c[y, x], format);
 	}
 
-	private static double[,] GetSubMatrix(Matrix matrix, int yOffset, int yLength, int xOffset, int xLength,
-		Func<Pixel, double> componentSelector)
+	private static float[,] GetSubMatrix(Matrix matrix, int yOffset, int yLength, int xOffset, int xLength,
+		Func<Pixel, float> componentSelector)
 	{
-		var result = new double[yLength, xLength];
+		var result = new float[yLength, xLength];
 		for (var j = 0; j < yLength; j++)
 		for (var i = 0; i < xLength; i++)
 			result[j, i] = componentSelector(matrix.Pixels[yOffset + j, xOffset + i]);
@@ -195,7 +195,7 @@ public class JpegProcessor : IJpegProcessor
 		};
 	}
 
-	private static byte[,] Quantize(double[,] channelFreqs)
+	private static byte[,] Quantize(float[,] channelFreqs)
 	{
 		for (var y = 0; y < DCTSize; y++)
 		{
@@ -207,7 +207,7 @@ public class JpegProcessor : IJpegProcessor
 		return _quantizedBuffer;
 	}
 
-	private static double[,] DeQuantize(byte[,] quantizedBytes)
+	private static float[,] DeQuantize(byte[,] quantizedBytes)
 	{
 		for (var y = 0; y < DCTSize; y++)
 		{
